Refactor CalloutAlert to its own module
